Move basic and wizard form into one class
