+++ src/Fields/Properties/WithWireModel.php
@@ -8,7 +8,7 @@ trait WithWireModel
     {
         $wireModel = $wireModel ?? $this->field->get('wire_model', 'change');
 
-        // "Defer" is Livewire's default so we don't want to return it as modifier.
+        /* "Defer" is Livewire's default so we don't want to return it as modifier. */
         return $wireModel !== 'defer' ? $wireModel : null;
     }
 }
