+++ src/Livewire/Concerns/WithComponent.php
@@ -13,8 +13,6 @@ trait WithComponent
     {
         $component = Str::replace('_', '-', $this->handle).'-form';
 
-        return Livewire::isDiscoverable($component)
-            ? $component
-            : "{$this->type}-form";
+        return Livewire::isDiscoverable($component) ? $component : "base-form";
     }
 }
