+++ src/Livewire/Concerns/WithTheme.php
@@ -17,17 +17,17 @@ trait WithTheme
 
     protected function theme(): string
     {
-        // Get the theme passed to the component in the view.
+        /* Get the theme passed to the component in the view. */
         if (isset($this->theme)) {
             return $this->theme;
         }
 
-        // Autoload the theme by form handle if it exists.
+        /* Autoload the theme by form handle if it exists. */
         if (ViewManager::themeExists($this->handle)) {
             return $this->handle;
         }
 
-        // Fall back to the default theme.
+        /* Fall back to the default theme. */
         return ViewManager::defaultTheme();
     }
 }
