+++ src/Livewire/Concerns/WithType.php
@@ -2,8 +2,6 @@
 
 namespace Aerni\LivewireForms\Livewire\Concerns;
 
-use Aerni\LivewireForms\Livewire\DynamicForm;
-use Aerni\LivewireForms\Livewire\WizardForm;
 use Livewire\Attributes\Locked;
 
 trait WithType
@@ -18,16 +16,14 @@ trait WithType
 
     protected function type(): string
     {
-        if ($this instanceof DynamicForm) {
-            return match ($this->type ?? null) {
-                'wizard' => 'wizard',
-                default => 'basic',
-            };
-        }
-
-        return match (true) {
-            ($this instanceof WizardForm) => 'wizard',
+        return match ($this->type ?? null) {
+            'wizard' => 'wizard',
             default => 'basic',
         };
+    }
+
+    protected function isWizardForm(): bool
+    {
+        return $this->type === 'wizard';
     }
 }
