+++ src/ServiceProvider.php
@@ -3,11 +3,10 @@
 namespace Aerni\LivewireForms;
 
 use Aerni\LivewireForms\Facades\Captcha;
-use Aerni\LivewireForms\Livewire\BasicForm;
+use Aerni\LivewireForms\Livewire\BaseForm;
 use Aerni\LivewireForms\Livewire\DynamicForm;
 use Aerni\LivewireForms\Livewire\Synthesizers\FieldSynth;
 use Aerni\LivewireForms\Livewire\Synthesizers\RuleSynth;
-use Aerni\LivewireForms\Livewire\WizardForm;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\Facades\Validator;
 use Livewire\Livewire;
@@ -60,8 +59,8 @@ class ServiceProvider extends AddonServiceProvider
     protected function bootLivewire(): self
     {
         Livewire::component('form', DynamicForm::class);
-        Livewire::component('basic-form', BasicForm::class);
-        Livewire::component('wizard-form', WizardForm::class);
+        Livewire::component('base-form', BaseForm::class);
+
         Livewire::propertySynthesizer(FieldSynth::class);
         Livewire::propertySynthesizer(RuleSynth::class);
